matrix change main

+++ tests/matrix_test/src/main.rs
@@ -29,7 +29,7 @@ fn main() {
 	let m: Matrix<u32> = Matrix(vec![vec![0, 0, 0, 0], vec![0, 0, 0, 0], vec![0, 0, 0, 0]]);
 	println!("{:?}", m);
 	println!("{:?}", Matrix::<i32>::identity(4));
-	println!("{:?}", Matrix::<i32>::zero(3, 4));
+	println!("{:?}", Matrix::<f64>::zero(3, 4));
 }
 
 #[test]
